Enable true Normal Mapping in the Tech Room.
For this feature to work, tangent maps must be pre-computed using XwaSmoother
and material files should have the new NormalMap setting enabled, like this:

NormalMap = Effects\AssaultGunboat.dat-0-0

The intensity of the effect is controlled by NMIntensity.

Also fixed a minor bug in the materials system: whenever a new material is
created I was previously loading the global material. This is wrong, it should
have used the ship's Default material instead.