Probable fix for the reticle pips misalignment.
See the cockpit look hook for more details. Here we're setting a flag to 1 if
we're in the Present 3D path and to 0 if we're in the Present 2D path (except if
we're in the Tech Room). This helps the cockpit look hook reset internal flags
that inhibit tracking when the reticle pips are being setup.

More testing is needed.