Fix Normal Mapping when normals are missing from a mesh.
For some reason, not all normals in a mesh are referenced when rendering. To
work around that, a counter has been added that keeps track of how many times
the tangent map has been computed. If we have tried more times than the maximum
number of meshes per OPT (256) then we create the tangent map with whatever we
have so far. Seems to work, but there may yet be some cases where it will break.