Minor bug fixes in shaders. Now we know the name of the OPT that is being loaded
PixelShaderAnim.hlsl is buggy: it's changing the material type when there's
transparency. For instance, the cracked glass in the T/F shows artifacts because
of this. Not sure how to fix this, but dividing fSSAOMaskVal by alpha seems to
fix this. Won't do it for now until I really understand what is supposed to
happen here.