Fix Raytracing performance bug.
The ray-tracer was using GPU Pro 3's code to do box-ray intersection tests, but
in that code, we did an intersection test in BVHIntersectBox(). If the test
failed, we returned T[0] = T[1] = -1. The caller then checks if T[1] >= T[0] to
see if there was an intersection, which would always be true because we were
setting both to -1 in BVHIntersectBox().

Removing that dumb test fixed the performance issues in the Tech Room. Now if
only I could get a nice BVH...