Fixed RTScale, and other bugs in the Raytracer.
The Tech Room changes the scale of the OPT, so the precomputed BVHs won't work
unless we have the right scale. To fix this, I added the mesh AABBs and mesh
vertex counts at the end of the .bvh files. That way we can check the AABBs of
the current mesh against the AABBs of the OPT and see the difference in scale.

Also moved the AABB class to LBVH.h. That's probably a better place for it.