Fix the OPT scale for Raytracing. Remove obsolete fields in the .bvh files.
Thanks to Jeremy, I no longer need the mesh AABBs nor the vertex counts. I can
read the scale used in the Tech Room directly off of XWA's heap. So let's remove
that code.

Also fixed the problem of using the wrong BVH when an OPT doesn't have a .bvh
file associated with it.