Performance fix: Add test that skips boxes that are behind the ray.
Lines are infinite, so the regular Tmax >= Tmin test will also catch boxes that
are behind the ray, making unnecessary tests and killing performance. To skip
those boxes, we need to add Tmax >= 0 to the check. We can't add Tmin >= 0
because rays originating inside a box will have Tmin < 0.

This little check makes a noticeable improvement in performance.