Workaround for the broken glass problem.
When broken glass is displayed, GLASS_MAT is used in ssaoMask.r. However, when
transparency is active, this material type will sometimes change after the
blending. To prevent that, I'm dividing by alpha. This isn't right, we should
not have a channel for materials, but instead just use one channel for
metallicity, one for glossiness, etc. So that blending won't change the type of
the material. This should work for now, though.