Ray-tracing with Embedded Geometry.
In order to support multiple OPTs and multiple mesh profiles per OPT, it might
be easier to have the geometry embedded in the BVH4 -- instead of having
additional SRVs for the vertices and indices. That way, we can just build a TLAS
referencing all the per-mesh BLASes without worrying about the vertex/index
offsets in each BLAS.

We still need to support TLAS traversal and per-BLAS offsets, but this is a step
towards that goal.