Fix/workaround for the MSAA edges bug.
The background returns a normal with alpha = 0. In that case, we bail out early
and return a shadeless color -- because the background shouldn't be shaded.
However, when MSAA is on, the alpha component is no longer 0 and it will vary
from 0 to 0.5. Increasing the "background shadeless margin" to 0.475 helps
reduce the edges, but there may be better fixes.