Probable fix for "lefover lighting" in the hangar.
When a mission is terminated by pressing H, we sometimes have "lefover" lighting
from the mission illuminating the hangar. This seems to happen because we're
forcing 2 lights in the hangar for a reason I forgot (but it seems to be related
to Hangar Shadow Mapping, see commit 19d7bd03). If I remove the line that forces
2 lights in the hangar, then shadow mapping is also gone. So, the "fix" here is
to keep 2 forced lights in the hangar, but also disable the second light by
setting its color to 0.