Expand Laser Lights to 16 slots. Add falloff and angle settings.
We can now have 16 dynamic lights, including laser lights and explosion lights.
Explosion lights now have inverse square attenuation and the reach is controlled
by a new setting called "falloff", the formula is:

attenuation = falloff / distance^2

So the bigger the falloff, the larger the area of effect, but it still will fade
with the square of the distance.

The following material settings were added:

	light_falloff (default 4)
	light_angle (default 0, meaning no angle)

The light_angle is not yet implemented, but it will control cone lights.