Add Object-Id to object-index map.
Jeremy's new d3d hook allows us to see the objectId associated with the current
texture with:

	scene->pObject->ObjectId

If we scan the *objects array, we can find which index in this array matches the
objectId. But to avoid scanning this array every time, an std::map
(g_objectIdToIndex) was added to cache the objectId-to-index mapping.