Add the Interdiction Effect.
Hyperjumps will now show a different effect when an Interdictor is pulling the
player out of hyperspace. We can only see the current region and mission index
from ddraw, so the Interdiction Map must be built using the origin region,
instead of the destination (which would be more logical).

Also fixed a few bugs: the twirl, and style were used by the ExplosionShader to
control other things, but the original values were not restored.

Populating the Interdiction Map will be done in a future commit. As it stands,
the effect cannot be used unless it's hard-coded in ddraw.