Add shake to the interdiction effect, single-line Regions
The shake effect can be modulated with the following Hyperspace.cfg settings:

	interdiction_shake_scale = 0.5 (default)
	interdiction_shake_scale_in_VR = 0.0 (default)

Set them to 0 to disable the shake effect. This parameter controls the speed of
the rotation:

	interdiction_angle_scale = 0.25 (default)

Regions are now specified in a single line, like this:

	[Interdiction]
	Region = 3, 2

The interdiction effect was delayed slightly to make the transition more
obvious.