Add the SkipWhenDisabled material property.
This property will force the current texture to be skipped when rendering as
soon as the ship instance becomes disabled. This is the first per-instance
material property added to ddraw.

To enable add the following to any .mat file section:

SkipWhenDisabled = 1

Also added a new property to the Material class: bInstanceMaterial. This
property must be set to true whenever a per-instance property is read from a
.mat file.