Add initial support for Instance Events/Animations.
Events and animations in .mat files can now be specified on a per-texture,
per-instance basis. Meaning that these events will only trigger for the specific
ship that is affected. For instance, adding the following to TieFighter.mat:

	anim_once = IEVT_DAMAGE_75, Effects\TieFighter.dat, (0-0,0.1,1)
	anim_once = IEVT_DAMAGE_50, Effects\TieFighter.dat, (0-1,0.1,1)
	anim_once = IEVT_DAMAGE_25, Effects\TieFighter.dat, (0-2,0.1,1)

Will cause damaged T/Fs to display damage animations when hit.

This is a work in progress. Currently all animations share the same slot in
g_AnimatedMaterials, meaning that they share the same timer. I still need to
enable per-instance animations with independent timers in
g_AnimatedInstMaterials, but this appears to be working nonetheless, and for
single-frame animations, the current code works just fine.