Add material Id, unique Instance Event Id and fix shield down events.
In order to properly instantiate Instance Events, we need to make a key for
instance events that combines the object Id and the material Id to make a new
64-bit key. Materials now have unique Ids to enable this. This makes it possible
to have independent animations per texture per object.

Fixed the shield events by adding a new independent event. However, this will
only work if shield down events are associated with the illumination layer.

For some reason, normal mapping gets de-activated when an animation is played.