Fix illegal access to std::map element in _tangentMap.
When computing the tangents for Normal Mapping, the current tangent map is
requested from _tangentMap.find() and then we update it. However, there's a case
where there may not be an entry in this map, we add it, but still use the
invalid iterator returned by find() to update it. This was causing an assert
when compiling ddraw in Debug mode.