Enable layered hull damage and shields down effects.
Animations can now be layered on top of other animations. There are two modes
that can be used, LYR_SCREEN and LYR_MULT. Use them like this:

	anim_once = IEVT_DAMAGE_75, LYR_MULT, Effects\GenericDamage.dat, (0-2,0.1,1)
	frame_seq = IEVT_SHIELDS_DOWN, LYR_SCREEN, Effects\GenericDamage.dat-2, 20, 0, 0.0, 0

The first line will add a hull damage layer and will mix it with the current
texture by multiplying it. The second will add the shields down effect and will
add it by screening it on top of the current texture.

This method enables generic damage textures to be used on a number of ships.