Fix the RenderEdgeDetector for missions without backgrounds.
Currently, the RenderEdgeDetector() effect is rendered on the first Execute()
call in Direct3DDevice. Usually, the Execute() is called at the top of a frame
because it's still used to render the background. However, missions without
backgrounds won't trigger this call, thus causing the CMD to be rendered without
wireframe.

A simple fix for this problem is to execute the edge effect right after the
DC foreground texture is resolved. This happens in PrimarySurface::Flip(), at
the end of the current frame. Thus, at the beginning of the next frame, we
should have the wireframe ready.