Revert Texture Lights to previous version. Tweak PBR shader settings.
The new texture lights are somewhat narrow and hard to see. Reverting to the old
version.

I noticed there was a bug in the PBR shader as well: the blackish term was a
bool, thus making a hard transition at this threshold. Setting that back to
float eliminated those edges.