Fix white edge around DC elements when MSAA is on.
If a DC element is displayed directly against the background it will display a
white edge around it if MSAA is on. This happens because the normal for the DC
element is getting w=1, when it should be SSAOAlpha instead (around 0.5 for
solid surfaces), as in the regular pixel shader. When w=1, the normal gets a
much brighter shade. This can easily be fixed by using SSAOAlpha.