Build the QBVH in one step instead of two.
In previous versions of the code, the QBVH was built using the Fast LBVH
algorithm by Apetrei, which builds a BVH2. This BVH2 was then converted into a
QBVH in a second step.

This version converts the BVH2 into a QBVH at the same time as the LBVH is
being built. It uses a custom greedy algorithm that saves about 2% memory space
but it appears to be faster (the fps for the T/F jumped from 29 to 47) because
the resulting tree is also flatter, so less checks are needed.

We still need an encoding phase, though. Future improvements can conceivably do
the encoding on-the-fly as well.