Enable single-step LQBVH build & encoding.
The QBVH is now built and encoded in a single step. As the Fast LBVH is building
the LBVH2, we collapse nodes to create QBVH nodes and then we encode them into
the buffer used by the shader. This produces a compact buffer that may have some
unused leading nodes. To make this buffer compatible with the current raytracer,
we're relabelling the offsets of the buffer so that the root starts at zero.

Further changes are needed to remove the re-labeling of the offsets (the
raytracer just needs to be aware that the root won't always be at offset 0).