Build the TLAS on every frame using RB trees and the camera frustrum.
The camera frustrum is captured every time OnSizeChanged() is called. Once it's
captured, the TLAS is built during frame construction through MainSceneHook().
Each time a mesh is about to be rendered, it's added to the TLAS using an RB
tree, so the resulting tree is balanced. The TLAS leaves will be OOBBs, so the
World View transform is also added to each leaf.

The RB tree still needs to be validated and encoded, but this appears to be
working and there doesn't seem to be a noticeable impact in performance.