Enable conditional display of textures based on MissionSetSpeed and throttle.
The new setting is called:

	DisplayIfMissionSetSpeedGE = MGLT

This refers to the MissionSetSpeed field which is set by the mission file
itself. This setting displays a texture if the ship is at full throttle and
the MissionSetSpeed field is greater than or equal to the value in the setting.
This helps display textures on ships that are following a flight path.