Fix orientation after hyperspace, collision and accel rates.
The gimbal lock fix is now disabled when in hyperspace. This ensures the
orientation provided by the game is not overwritten.

Increased the precision of the InitYPR() function, that makes the collisions
less jerky.

Reduced accel rates as they seem to match better the original values used by the
game.