Reallocate RT buffers when the number of total nodes increases.
Yet another step towards in-flight RT. In this approach, we're using a large
buffer to hold all the nodes from all the BLASes as needed. Whenever the total
nodes increases (for instance, after seeing a new mesh), we need to release the
RT buffers, create new ones and repopulate them.

Oh, and we can't store the WorldView matrices in the g_LBVH map: the matrices
change on every frame!