Add TLAS leaf node encoder, FastLBVH main code is now using templates.
The TLASLeafItem and BLAS LeafItem are both std::tuples and they share the same
data in the first 3 entries. The code that builds the FastLBVH can be converted
to use templates so that it works for both the TLAS and BLAS.

The TLAS leaf node is missing the matrix slot index. So the code needs to be
further refactored to get the slot from the g_LBVHMap.