Add g_TLASMap, update the BLAS and TLAS map in the same place.
The matrixSlot is now shared between the TLAS leaves and the BLASes. To build
the TLAS, I'm now using a TLAS map that maps meshKeys to matrixSlots.

The TLAS map gets cleared at the beginning of every frame.

The BLAS map only gets cleared on OnSizeChanged().

The hypothesis here is that BLASes only get built when we first see them. Their
matrices change on every frame, and the matrices for the current frame are
indexed by the TLAS map. The matrixSlots are then copied from the TLAS map into
the BLAS map when the relevant mesh is rendered.