The TLAS map needs to use the centroid as part of the key. More refactoring.
There's a bug somewhere in the single-step QBVH builder that sometimes causes
the encoding index to become negative. This happens when there's 7-10 TLAS
leaves on the screen, but only sometimes. Since I can't get a repro, I just
increased the number of nodes in CalcNumNodesQBVH(). Not ideal, but let's move
on.

The TLAS map needs the mesh and the centroid: the same mesh may appear multiple
times in different elements of the Flight Group, for instance. We need to have
a single BLAS for all of them, but multiple matrices and multiple TLAS leaves.

The BLAS leaf item, on the other hand, can't store a matrix, for the same
reason: the same mesh may be used in multiple places.