The matrices need to be transposed before sending them to the shaders.
Not sure why the matrices need to be transposed... the Tech Globe didn't need
that, but for the in-flight raytracer that's what we need to do.

Also added more debug code to verify the TLAS tree (it's fine). The TLAS tree
stores both the AABB and OBB and both tests pass properly.