Fix for the Single-step Fast QBVH builder.
We should not convert BVH2 nodes twice. The root sometimes gets converted twice
because I added a special case for that in the code. I've put some checks to
prevent this, and hopefully that'll fix the random crashes I've been seeing.

Also changed the BLAS builder to the 3-step builder because that's a safer
approach.