Refactor code to allow coalesced BVHs.
Added a flag that can coalesce all the FaceGroups that belong to a mesh. This
flag is set to false for all FGs, but we can set it to true for meshes like the
cockpit that have no LOD to create trees with better performance.

Also refactored the code so that all BLASData entries have a FaceGroup map. For
non-coalesced BVHs, this map only has a single entry.