Ancient bug: Reset XWA Lights when exiting the hangar or ShadowMap will not work
If we exit the hangar in external view, then shadow mapping will stop working
when lights are faded. The reason is that for ShadowMapping, we intentionally
set light sm_black_level[0] = 1 when rendering hangar shadows. We need to do
this to avoid cockpit shadow mapping when there's no cockpit OBJ.

However, once we exit the hangar, that light stays at 1.0 if we're in external
view. If it also happens to be the slot for the Sun, then there will be no
active lights after all lights have been tagged. The fix is to reset the shadow
mapping params every time we exit the hangar.

Also fixed some missing Annotation events in these areas.

Also added checks to prevent tagging lights if we're in the hangar.

Also allowed tagging lights in external view as long as y_center has been
computed. I'm not sure this is a good idea, but let's try it.