Revert light tagging in external view, fix ShadowMap bug in hangar.
The hangar must be allowed to do shadow mapping when RT cockpit shadows is on.
The name of the variable has been renamed accordingly and the condition to
inhibit shadow mapping if RT is on is now more nuanced.

Also, it looks like tagging lights while in external camera mode brings many
problems. I need to revisit that later. Tagging lights while in external
camera mode is now disabled again.