Fix shadows for the PBR shaders, tag lights in external view.
Still undecided whether tagging lights in external view is a good idea or not,
but after fixing some lighting bugs, it appears to be working OK.

The main lighting loop now has two cases: for shadow casters, we use the
total_shadow_factor, which includes both RT and SM and for non-casters we avoid
that factor. That eliminates terminator-related artifacts.