Fix hangar shadows when the PBR shader is on.
The PBR shader was modified to handle shadow casters separately to fix some
artifacts. This caused a problem when the external view is used in the hangar:
the first light is disabled, causing everything to go dark.

I added an exception to this rule so that the first shadow caster remains active