Build g_FGToSetMap from the data in OptHeader during D3dRendererOptNodeHook()
D3dRendererOptNodeHook() is called multiple times per frame, and it comes with
an OptHeader pointer that can be used to inspect all the OptNodes in the OPT
being rendered.

This is how we can see which FaceGroups belong to a separate LOD (or group, in
general terms). This information can be used to build a FaceGroup to Set map.
We can use that map to coalesce meshes when building a BVH.