Coalesce all FaceGroups by their LODs and improve performance.
Thanks to Jeremy, I can now parse the OPT nodes and figure out which face groups
belong to the same LOD. With this information, we can now cut down on the number
of TLAS leaves and make better BLASes, improving performance.

The key to the BLAS map is now a meshId-LOD pair, as it should be.