Reset the Scissor Rect when rendering RTShadowMask.
The scissor rect was causing the RTShadowMask to become clipped when all GUI
elements where turned off. We need to save/restore the scissor rect along with
the rest of the context and then we need to reset it to full screen for all
post-processing.

Also minor code cleanup in various places.