Simply Tangent Map computation by parsing OptNodes and using all LODs.
In order to properly compute tangent maps, we need all the Face Groups belonging
to all the LODs in the current mesh. To do this, we need to parse the OptNode
when the OPT is being rendered and we build a mesh to FGs map (g_meshToFGMap)
with this information.

This map is later used to find all the FGs for the current mesh when computing
the tangent map. This greatly simplifies the logic since we only need to retry
the tangent map computation when the map isn't complete. Once the map complete,
we compute all the tangents in one try.