Slightly increase the shadow terminator to avoid artifacts on tangential lights
When a light source is tangential to a surface, it will sometimes flicker as it
goes in and out of the shadow zone. To avoid this, the test that checks if a
surface is facing away from a light source is now slightly larger.

I'm not sure this is a good idea, but it appears to help in some cases (i.e. the
Rebel Platform).