Parse RANDOM_LOC and SCALE damage texture modifiers.
There's a fair bit of refactoring in this commit and these modifiers are not yet
active (and therefore, they're not tested either).

These damage texture modifiers will help us place damage randomly on a surface
and scale it as necessary so that we can add damage to both capital ships and
fighters.