UpdateBVHMaps() should only be called when RT is enabled.
The condition to call UpdateBVHMaps() starts with:

if ((g_bRTEnabledInTechRoom || g_bRTEnabled) && ...

However, g_bRTEnabledInTechRoom is true by default, meaning that this section
will be true even if Raytracing is disabled during regular flight. In other
words, UpdateBVHMaps() is called for regular flight even if RT is off. This will
update the std::maps needed to build the BLAS, but the code that consumes that
map isn't active. Hilarity may ensue.

This condition should read:

if (((g_bRTEnabledInTechRoom && g_bInTechRoom) || g_bRTEnabled)

to ensure that this block is called when RT is enabled if we're in the Tech Room
or if we're in regular flight and RT is enabled.