Fix dark screen bug when exterior view is enabled, remove crutch.
The crutch I'm referring to is this one:

https://github.com/Prof-Butts/xwa_ddraw_d3d11/commit/f19db50843d726737512fad514f9c5a284a3389b

The problem came down to the g_bTargetCompDrawn flag. The new d3dhook makes the
code jump from Execute() to the EffectsRenderer, where this flag isn't updated.
So the fix is to update this flag in EffectsRenderer::DoStateManagement(). That
fixes the state and errors that came from the mistmatch.