Fix small artifacts in the beam weapon DC area when no beam weapon is equipped.
If a ship doesn't have a beam weapon equipped, and no cover texture is specified
in the DC file either, then the code will still replace the textures anyway and
it will display a small upside-down version of the DC foreground texture. This
shows as artifacts on the beam weapon area in the cockpit.

A quick fix here is to avoid replacing the DC textures if there's nothing to
display (i.e. the beam weapon is not equipped) and no cover texture exists
either.