Properly initialize the cockpit instrument status.
Instead of using the initial state of the cockpit instruments, I was using the
default constructor CockpitInstrumentState() as initial state. The problem is
that this constructor sets all the instruments to true -- even if the current
ship isn't equipped with all the relevant devices. As a result, ships without
the beam weapon or shields would falsely trigger the damage events for these
instruments.

I'm now using the initial state of the instruments to initialize the cockpit
state. This no longer triggers the damage events for ships that are not equipped
with beams or shields.