(Disabled) Prevent Black Suns.
Some international languages may cause suns to get a black color if the FGs are
messed in the xwa.tab file. This code prevents such black suns, but it's
currently disabled because the proper fix is to have a *good* xwa.tab. I leave
this (disabled) code here just in case we ever need it again (but in that case
it should also come with a chicken switch, probably in SSAO.cfg).

For testing purposes, install the German language and try the second region of
B5M3.