Simplify the inner loop of the DirectBVH builder.
We don't need to get the min/max of boxL and boxR for every primitive. We can
use the split point and the input AABB instead. This leads to a slightly less
efficient loop (since a few more iterations may be needed if we end up with 0
leaves on one side), but in a GPU implementation this is important because the
min/max boxes are computed with global atomics (which don't work on floats
anyway) and that destroys the performance.

This proves that it is possible to build the tree without using these global
atomics -- even if we do need to keep track of countL, countR.