Get rid of std::tuple and replace with regular structs.
std::tuple is nice, but I get confused keeping track of all the indices. It's
probably better to have named fields -- as in a struct -- and use inheritance so
that TLASLeafItem's get the same fields as a regular LeafItem plus the extra
two that it needs.

Also, maybe I can get rid of the template functions now...