Fix the DirectBVH2GPU builder (it needs a few more iterations).
The max number of iterations isn't log2(numPrimitives) -- that's the best case
scenario only, when each split halves the number of primitives on each side.
Reality is different: trees aren't perfectly balanced and that's reflected on
the number of iterations. The worst-case is iterating once per inner node.

This code appears to work fine, but more testing is needed.