PLOC: Replace vector distance with AABB surface area.
The original PLOC paper doesn't use euclidean distance between centroids, it
uses the surface area of the joint AABBs. So let's do that too.

Also removed the std::vector used to store the clusters and replaced it with two
regular arrays. This is probably faster and that's how we would implement this
algorithm in a GPU anyway.