Start decoupling inner node indices from QBVH inner node indices.
The current encoding scheme encodes inner nodes starting at index 0 and leaves
starting at a fixed offset. However, not all inner nodes will be filled after
the BVH is built, leaving a gap between the last inner node and the first leaf
node. This isn't great because if all nodes were contiguous we could use the
cache when traversing the tree. Right now, there's going to be a cache miss
whenever we jump from an inner node to a leaf node.

To prevent this, we need to decouple the encode index from the node index. Then
we need to have an encode index and a "scratch buffer" index. More work is
needed and this is WIP still.