Refactor the EmitLeaf code to decouple the scratch and buffer indices.
Also modified the LeafEncode function to encode the index of the LeafItem
instead of the primitive's index. This helps when refitting the DBVH4 because
we can use the ref field to index the LeafItem and fetch the relevant AABB.
The number of inner nodes is no longer used to figure out when a node is a leaf.