Encode leaves inline (leaves are no longer pre-encoded before building a BVH).
In order to avoid cache misses, I'm now encoding the leaves wherever they are
first attached to the BVH. In other words, there's no longer an inner node
section and a leaf section. The encoding can be improved by encoding all 4
children of a node contiguously, but that'll have to wait.

Also, there's some code cleanup needed here. The previous idea where I had a
separate scratch index and a buffer index was too complicated. The index is now
the same for the scratch and the buffer, but I need to coalesce the indices
again.