Use an inner node index map to save space.
The previous approach that has a 1:1 correspondence between scratch nodes and
buffer nodes is very wasteful because there are lots of inner nodes that
correspond to leaf nodes that won't be used. To remove most of that waste, we
now need a way to map encode indices to inner node indices. One such way would
be an std::map<EncodeIndex, InnerNodeIndex>, but that isn't GPU-friendly, so
here we can use an integer map. I can use an std::map later too, that's why I
wrapped everything inside inline functions.