Replace the AVL builder with a balanced online builder.
The balanced builder is based on Erin Catto's presentation on the topic, but it
also has an alternate balancing scheme based on:

https://github.com/briannoyama/bvh

which is in turn based on:

Fast, Effective BVH Updates for Animated Scenes by Kopta et al.

The first balancing scheme works kind of alright, but the tree is not great and
is outperformed by the LBVH. The second scheme doesn't really work for me, but
I didn't bother doing a full conversion from golang (I don't think balacing
based on depth is a great idea anyway).

The first approach has more future, but the balancing scheme is different from
the paper and it needs an initial build that is high quality. So maybe this can
be used to update TLASes and trigger a full rebuild only after N frames.