Enable the TLAS path to use the DirectBVH builder.
The DirectBVH builds faster and uses less resources than the FastQBVH. The
performance is about the same.

Also did a minor fix for the DirectBVH builder: we should never access the
encode index from a leaf node. This doesn't make sense (only inner nodes should
be accessed this way) and in the TLAS path, the leaf nodes store additional data
anyway.

Also changed the BuilderType enum for an enum class... it's more classy that
way. Look at that, I just made a programming funny.