Add code to compact BVH trees, and code to dump TLAS input geometry.
Turns out the DirectBVH builder does not produce compact QBVH trees, so we need
an additional step to save memory. This is what was causing the crashes when
this builder was used to build the TLAS: sometimes there's more than
(2/3 * numPrimitives) inner nodes. In fact, I think it can degenerate to
(numPrimitives - 1) inner nodes in the worst-case scenario -- just like a
regular BVH2. Bummer.