Add the CompactBVH() to the TLAS build path for the DirectBVH builder.
This effectively creates smaller trees, but it's using a lot of scratch memory,
so it's not great either. I need to find a better solution.

Also, I refactored the DirectBVH code to use a common state held in a struct.
I could've used a class, but I'm not sure I can easily jump into arbitrary
functions from threads if they are class members.