Allow split reprocessing to prevent overflowing the number of inner nodes.
This code reprocesses primitives when there's 2 or more empty slots in an inner
node. The splits are recomputed to hopefully reduce the number of empty slots
the second time around. This saves some inner nodes at the expense of a few more
iterations. Still not sure if this will be enough to prevent overflows, though.