Add code to compute Eigensystems.
... for symmetrical covariance matrices. Because OBBs!

Original code from:

https://www.mpi-hd.mpg.de/personalhomes/globes/3x3/index.html

Now here's a little fact: jokes have eigensystems too. The joke may change some
details -- or even language -- but there's always a part of the joke that stays
the same, no matter what. This is the joke's eigensystem! And once you see it...
the joke is ruined forever. You're welcome!