Partial fix for the miniature display depth buffer bug.
The depth in the miniature CMD display was wrong because the depth stencil was
being set to NULL. This happened back when we switched to Instanced Rendering.
To fix the display in pancake mode, we just need to put the depth stencil back
in.

However, in SteamVR mode, doing this causes a blank display and other artifacts,
so I'm keeping the NULL depth stencil in that case. I thought this would be
fixed by using DrawIndexedInstanced with 1 instance, but it didn't work. So the
depth is still messed up in VR for now.