Fix crash in TFTC, B8M5-R.
This crash was happening because for some reason, release 2.4.6 had the DBVH4 as
the default builder and it's still WIP. B8M5-R has an interesting configuration
for the TLAS when exiting the hangar from camera view 2. This TLAS has a very
elongated box with all primitives at either side of it. The regular split
selected the same axis twice, creating a node with 2 empty children. This
repeated a couple of times and we overflowed the number of inner nodes.

To fix this, I've switched to using full boxes instead of ranges, and I'm also
adding a new step that merges slots with exactly 2 children into a single node.
This appears to have avoided the problem, but there may be more corner cases
that will need to be handled later.