Add support for DATReader 1.0.3 and for legacy and future DAT files.
Legacy DAT files (i.e. all the DAT files published up to this point) need their
images flipped and channels swizzled; but moving forward this won't be necessary

DATReader 1.0.3 supports both legacy and future formats. However, ddraw needs to
decide when to flip and swizzle the images. To this effect, this ddraw will flip
the images by default, unless instructed not to by the following new setting in
ddraw.cfg:

	FlipDATImages = 0 (default = 1)

For the current XWAU this is not necessary because all public DAT files need to
be flipped. Future releases will need this until I change the default setting.