Fix missing map icons.
Map icons are now identified so that they can be processed differently. For some
reason, their rhw component is "inverted". I believe the real cause is that the
game flips the Depth Comparison function, but I fail to catch that event. Either
way, flipping the rhw value also works. Plus it looks like the input.color needs
to be used for the map colors and that is controlled by fDisableDiffuse property
anyway, which needs to be unset for map icons -- so we need to detect them.