Prevent crashes in the Tech Room.
In a previous commit, I moved g_meshToFGMap.clear() to OnSizeChanged(). I did
this to fix disappearing normal maps after pressing H. However, this had the
side effect of not clearing this map when new OPTs are loaded in the Tech Room.
Since FGs were still cached but with wrong vertex counts this caused random
crashes. One easy way to get this crash is to load the first 28 ships in the
Tech Room by just pressing "next". Eventually, when the Escort Shuttle is loaded
the crash happens when computing the normal maps for this ship.

The fix here is to clear g_meshToFGMap when new OPTs are loaded in the Tech
Room, in D3dRendererOptLoadHook().