Fix normal mapping when greebles are activated.
Not sure why, but the Z component was flipped and that was causing some serious
illumination artifacts. I must've had a good reason to do that, but now we don't
need it anymore.

For the record, I just copied what I'm doing for the regular normal mapping path
in XwaD3dPixelShader.hlsl, in the "if (bDoNormalMapping)" code block.