Fix explosion washout.
Explosions don't apply any material properties. This causes the background mats
to be blended with the foreground explosion making them look "washed out".

To fix this, I've enabled a special control mask to allow explosions to blend
their material properties so that they look solid when rendered in the
foreground.