Enable simple BLAS intersection test and display ray origin and hit.
Another step in the process of restoring the Active Cockpit. Here the code loops
through all the AC TLAS leaves (i.e. all the cockpit BLASes) and selects the
best intersection but it doesn't iterate each BLAS to find the best primitive.
Still, both the cursor and the intersection are displayed, so... progress!