Remove tangent map buffers. Use cotangent_frame() instead.
Turns out the tangent map buffers were causing crashes in the hangar (and
probably other areas). Apparently this was caused by OPTs with profiles as that
may cause the mesh buffers to be swapped at runtime and the number of vertices
no longer matches. Also, the hangar reloads the mesh buffers every time a new
view is displayed, causing all tangent maps to be recomputed (and probably
leaked as well). So, let's just do the tangent maps in HLSL now.