Emulate joystick roll and add ESC AC command.
We need to be able to exit the game during flight without touching the keyboard,
so let's add the ESC command.

Added support for joystick roll by twisting the VR controllers. Yaw and pitch
are computed again by using angles. The yaw and pitch come from quaternions, but
the pitch is a simple projection to the Y-Z plane. This seems to work, but it
needs a somewhat large deadzone.