Enable AC when Raytracing is off.
AC now uses BVHs to quickly test intersections in some areas. But if Raytracing
is off, then no BVHs are built. We need to enable some RT paths when AC is on to
make sure BVHs are still built even if RT is off -- but we don't want to render
shadows in this case!