Enable region highlighting for VR geometry.
Specific regions can now be highlighted in the VR keyboard. This is useful for
sticky keys, like Ctrl, Alt, Shift.

I also added code to allow using the VR keyboard with a single controller (it
previusly defaulted to using the pose from the controller that didn't trigger
the keyboard, which would fail if only one controller is present).