Enable closest-point-on-triangle AC activation. Fix LaserPointerCommon debug.
The debug code in LaserPointerCommon has been fixed. It will now display the
triangles properly.

Fixed the problem that prevented using orthogonal distance to the VR keyb (I
just had to keep iterating and looking for the best distance).

Enabled closest-point-on-triangle test for regular AC elements. This makes the
glove "snap" to the closest active AC element. This works so well that I can
even activate the S-Foils on the X-Wing without even looking (yes, muscle memory
is now a thing in VR).

TODO: Enable closest-point-on-triangle for the VR gloves themselves (this
involves modifying the BVH traversal code as well).