Snap to the nearest key in the VR keyboard. Fix bug that triggers Point profile
The intersection with the VR keyboard is now snapped to the nearest key. That
avoids the small gaps between keys.

The point profile was being triggered by a hard-coded 0.18 value in the code.
It should be GLOVE_NEAR_THRESHOLD_METERS instead to be consistent with other
uses of this threshold.