Display the VR gloves properly when the Gunner Turret is active (WIP).
The gunne turret is weird. It doesn't spin around the POV, so it's hard to align
it properly with the VR gloves. So, instead, let's keep the VR gloves in
view-space coords (i.e. their original position coming from SteamVR) and let's
make it work with the keyboard.

The gloves are displayed properly after this commit, but the VR keyboard is
useless in the gunner turret at this point.